+++ src/app/(main)/(posting)/category/[id]/page.tsx
@@ -4,7 +4,7 @@ import styles from '@app/(main)/(posting)/category/[id]/page.module.css';
 import ArticleBox from '@/common/component/Article/ArticleBox';
 import { articleList } from '@/common/component/Article/Article.mock';
 import { articleType } from '@/common/component/Article/ArticleType';
-import ArticleListControls from '@/common/component/ArticleListControls/ArticleListControls';
+import ArticleListControls from '@/common/component/Article/ArticleListControls';
 
 const CategoryPage: FC<IDPropsType> = ({ params }) => {
 	return (
